Fix action button counter glitch in most scenarios
Adds monospacedDigit modifier to the counter next to action button counter labels. This fixes the layout glitch that happens when the counter changes, most noticeably when you tap the boost or favorite button except for when the counter gains or loses a digit (9 -> 10, 99 -> 100 etc…)
Still greatly improves the actions view experience though.