Don't show the source-language before translating (#1008)
* Clear up the translate-button

The Translate button on the post and in the context menu now does not show the
source language to avoid user confusion if the language set by the poster is
not the actual language of the post. This language is now only used to decide
whether to display the button in the main view. The Translate button in the
context menu is independent.

Signed-off-by: Paul Schuetz <pa.schuetz@web.de>

* Show the translation-source-language

The source-language of a translation is now shown.

Signed-off-by: Paul Schuetz <pa.schuetz@web.de>

* Remove unused strings

The now unused translate-from-strings are removed.

Signed-off-by: Paul Schuetz <pa.schuetz@web.de>

---------

Signed-off-by: Paul Schuetz <pa.schuetz@web.de>
Co-authored-by: Thomas Ricouard <ricouard77@gmail.com>