Add option to not show replies
An user option to not show replies. There is the possibility to overwrite the option via a flag set when creating the view. This is necessary to show the replies when being in the specific tab in the account view. This option is triggered for all statuses in the account views, the display of replies is there (account tab) only managed via the old way.
Closes #1209

Signed-off-by: Paul Schuetz <pa.schuetz@web.de>