Added a "Continues a thread" label to status rows
Just like "boosted by" and "replied to", I do think that thread label brings value when seeing a message that might be out of context in the timeline. This helps bringing context without having to press the message to check if there is a thread above.
Pressing the message will do the same and allow to consult the whole thread at a glance